% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_estimateLambda.R
\name{calcLambda}
\alias{calcLambda}
\title{Estimate the regularization parameter}
\usage{
calcLambda(object, model, seq_lambda1, data.fit, data.test, warmUp = FALSE,
  keep.fit = FALSE, refit.pLVM = TRUE, fit = "BIC", order = "lambda1",
  trace = TRUE, ...)
}
\arguments{
\item{object}{the penalized latent variable}

\item{model}{the non penalized latent variable}

\item{data.fit}{the data used to fit the model}

\item{data.test}{the data used to test the model}

\item{warmUp}{should the new model be initialized with the solution of the previous model (i.e. estimated for a different penalization parameter)}

\item{keep.fit}{should the penalized LVM be exported}

\item{refit.pLVM}{should the penalized LVM be fitted on only the non 0 parameters}

\item{fit}{criterion to decide of the optimal model to retain among the penalized models.}

\item{order}{the order of the sequence of penalties: according "lambda1" or "lambda1.abs"}

\item{trace}{shoud the execution of the function be traced}

\item{...}{additional arguments - e.g. control argument for estimate.lvm}

\item{seq_lambda}{the sequence of penalisation paramaters to be used to define the sub model}
}
\description{
Find the optimal regularization parameter according to a fit criterion
}
\examples{


}

