% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lava_reducedMoment.R
\name{calcLP.lvm}
\alias{calcLP.lvm}
\title{Compute the linear predictor}
\usage{
calcLP.lvm(x, p, data, lp.x, lp.link, lp.endo)
}
\description{
Compute the value of the linear predictors of a LVM and store it into the dataset
}

