% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lava_link.R
\name{cancel}
\alias{cancel}
\alias{cancel.plvm}
\title{Remove a new link between two variables in a plvm model}
\usage{
\method{cancel}{plvm}(x, value, clean = TRUE, ...)
}
\arguments{
\item{x}{a lvm model}

\item{value}{the names of the links that should be removed}

\item{clean}{should the lvm object be simplified using the \code{clean} function}

\item{...}{other argument to be passed to \code{clean}.}
}
\description{
Generic interface to remove a link in a plvm.
}
\examples{

#### penalized lvm ###
m <- lvm(Y ~ X1 + X2 + X3)
pm <- penalize(m)
cancel(pm, Y~X1+X2)
cancel(pm, Y~X1+X2+X3)

}

