% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lava_reducedObject.R
\name{cancelLP}
\alias{cancelLP}
\alias{cancelLP<-}
\title{Remove variable from the linear predictor}
\usage{
cancelLP(x, ...)

cancelLP(x, ...) <- value
}
\arguments{
\item{x}{\code{lvm}-object}

\item{lp}{should the name of the variables corresponding to the linear predictors be returned?}

\item{expar}{should the external parameters be also removed}
}
\description{
Remove one or several variable from the linear predictor
}
\examples{

## regresssion
m <- lvm()
m <- regression(m, x=paste0("x",1:10),y="y", reduce = TRUE)
vars(m)

cancelLP(m)
cancelLP(m, lp = "LPy", link = c("y~x1","y~x2"))

## with penalization
m <- lvm()
m <- regression(m, x=paste0("x",1:10),y="y")
pm <- penalize(m, reduce = TRUE)
cancelLP(pm)
coef(cancelLP(pm))

}

