% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_penalize.R
\name{cancelPenalty}
\alias{cancelPenalty}
\title{Remove penalty from a penalized latent variable model}
\usage{
cancelPenalty(x, ...)
}
\arguments{
\item{x}{\code{plvm}-object}

\item{link}{the penalty that should be removed}

\item{value}{the penalty that should be removed}

\item{simplify}{if the object contain no more penalty should it be converted to a non penalized lvm}
}
\description{
Remove one or several penalties from a penalized latent variable model
}
\examples{
m <- lvm()
m <- regression(m, x=paste0("x",1:10),y="y")
pm <- penalize(m)

cancelPenalty(pm, link = "y~x5")
cancelPenalty(pm) <- "y~x1"
cancelPenalty(pm) <- y~x1
}

