% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lava_extract.R
\name{checkData}
\alias{checkData}
\alias{checkData.lvm}
\title{Check that validity of the dataset}
\usage{
checkData(x, data, ...)

\method{checkData}{lvm}(x, data, ...)
}
\arguments{
\item{x}{a lvm model}

\item{data}{the dataset containing the variables used to estimate the lvm.}

\item{...}{not used}
}
\description{
Check the validity of the dataset used to estimate a lvm.
}
\examples{
m <- lvm()
regression(m) <- c(y1,y2,y3)~u
regression(m) <- u~x
latent(m) <- ~u

d <- sim(m,1e2)

try(checkData(m, d)) # return an error

checkData(m, d[,-4])
}

