% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lava_extract.R
\name{checkVar}
\alias{checkVar}
\alias{checkVar.lvm}
\title{Extract the name or the position of the variance coefficients}
\usage{
checkVar(x, data, ...)

\method{checkVar}{lvm}(x, data)
}
\arguments{
\item{x}{a lvm model}

\item{value}{should the name of the coefficient be returned? Else return the coefficients}
}
\examples{
m <- lvm()
regression(m) <- c(y1,y2,y3)~u
regression(m) <- u~x
latent(m) <- ~u

d <- sim(m,1e2)
checkVar(m, d)

checkVar(m, d[,-4])
}

