% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_extractLVM.R
\name{coef0.plvmfit}
\alias{coef0.plvmfit}
\title{Extract Model Coefficients}
\usage{
\method{coef0}{plvmfit}(x, tol = 0, operator = "<=", penalized = FALSE,
  value = TRUE)
}
\arguments{
\item{x}{a penalized lvm model}

\item{tol}{the threshold below which a coefficient is considered to be null}

\item{operator}{the operator used to extract the coefficients}

\item{penalized}{should only the penalized coefficient be returned}

\item{value}{should the value of the coefficients be returned? Otherwise their position.}
}
\description{
Extract the 0 or non 0 coefficients
}
\examples{
set.seed(10)
n <- 300
formula.lvm <- as.formula(paste0("Y~",paste(paste0("X",1:12), collapse = "+")))
mSim <- lvm(formula.lvm)
df.data <- sim(mSim,n)

pm <- penalize(mSim, c("Y~X1","Y~X4", "Y~X10"))
pm.fit <- estimate(pm, lambda1 = 1e5, data = df.data, control = list(constrain = TRUE))
coef0(pm.fit)
coef0(pm.fit, operator = ">")
coef0(pm.fit, operator = ">", penalized = TRUE)

}

