% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lava_extract.R
\name{coefCov.lvm}
\alias{coefCov.lvm}
\title{Extract the name or the position of the covariance coefficients}
\usage{
\method{coefCov}{lvm}(x, value = FALSE)
}
\arguments{
\item{x}{a lvm model}

\item{value}{should the name of the coefficient be returned? Else return the coefficients}
}
\examples{
m <- lvm()
regression(m) <- c(y1,y2,y3)~u
regression(m) <- u~x
latent(m) <- ~u

coefCov(m)
coefCov(m, value = TRUE)

covariance(m) <- y1 ~ y2
coefCov(m)
coefCov(m, value = TRUE)

}

