% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{combine.formula}
\alias{combine.formula}
\title{Combine formula}
\usage{
combine.formula(ls.formula, as.formula = TRUE, as.unique = FALSE)
}
\arguments{
\item{ls.formula}{a list of formula}

\item{as.formula}{should as.formula be applied to each element of the list}
}
\description{
Combine formula by outcome
}
\examples{
\dontrun{#' 
combine.formula(list(Y~X1,Y~X3+X5,Y1~X2))
lava.options(symbol = c("~",","))
combine.formula(list("Y~X1","Y~X3+X5","Y1~X2"))
lava.options(symbol = c("<-","<->"))
combine.formula(list("Y<-X1","Y<-X3+X5","Y1<-X2"))

combine.formula(list(Y~X1,Y~X3+X1,Y1~X2))
combine.formula(list(Y~X1,Y~X3+X1,Y1~X2), as.unique = TRUE)
}
}

