% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_proxOperator.R
\name{composeOperator}
\alias{composeOperator}
\title{Composition of several proximal operators}
\usage{
composeOperator(...)
}
\arguments{
\item{...}{proximal operators}
}
\value{
a function
}
\examples{
lava.penalty:::composeOperator(proxL1, proxL2)

}

