% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lava_cv.R
\name{cvCheck.lvm}
\alias{cvCheck.lvm}
\title{Test the sensibility of the lvm estimate to the initialization points}
\usage{
\method{cvCheck}{lvm}(object, data, factor.vcov = 9, n.init = 100,
  ncpus = 1, verbose = TRUE, ...)
}
\arguments{
\item{object}{a lvm model}

\item{data}{a data frame}

\item{factor.vcov}{inflation factor for the variance when sampling the initialization points}

\item{n.init}{number of initialization points to be used}

\item{ncpus}{the number of CPU to be used}

\item{verbose}{should a progression bar be displayed?}

\item{...}{additional arguments to be passed to estimate}
}
\value{
a data frame/cvlvm object containing the convergence status (by default 0 indicates successful convergence, see ?optim), the value of the log-likelihood and the estimated parameters (in columns) for each initialization (in rows)
}
\details{
Simulation is based on a multivariate truncated normal law (even though it is not satifying for the variance components)
}
\examples{
#' 
m <- lvm(list(y~v1+v2+v3+v4,c(v1,v2,v3,v4)~x))
covariance(m) <- v1~v2+v3+v4
dd <- sim(m,10000) ## Simulate 10000 observations from model
e <- estimate(m, dd) ## Estimate parameters

system.time(
summary(cvCheck(m, dd))
)

system.time(
summary(cvCheck(m, dd, ncpus = 4))
)
}

