% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lava_cv.R
\name{cvCheck}
\alias{cvCheck}
\alias{cvCheck.lvm}
\title{Test the sensibility of the lvm estimate to the initialization points}
\usage{
cvCheck(object, ...)

\method{cvCheck}{lvm}(object, data, factor.vcov = 1, n.init = 100,
  keep.cov = TRUE, ncpus = 1, trace = TRUE, ...)
}
\arguments{
\item{object}{a lvm model}

\item{...}{additional arguments to be passed to estimate}

\item{data}{a data frame}

\item{factor.vcov}{inflation factor for the variance when sampling the initialization points}

\item{n.init}{number of initialization points to be used}

\item{keep.cov}{should the covariance between parameter be kept to simulate the initialization points}

\item{ncpus}{the number of CPU to be used}

\item{trace}{should a progression bar be displayed?}
}
\value{
a data frame/cvlvm object containing the convergence status (by default 0 indicates successful convergence, see ?optim), the value of the log-likelihood and the estimated parameters (in columns) for each initialization (in rows)
}
\description{
Test the sensibility of the lvm estimate to the initialization points
}
\details{
Simulation is based on a multivariate truncated normal law (even though it is not satifying for the variance components)
}
\examples{
m <- lvm(list(y~v1+v2+v3+v4,c(v1,v2,v3,v4)~x))
covariance(m) <- v1~v2+v3+v4
dd <- sim(m,10000) ## Simulate 10000 observations from model
e <- estimate(m, dd) ## Estimate parameters

\dontrun{
summary(cvCheck(m, dd, ncpus = 1))
summary(cvCheck(m, dd, ncpus = 4))
}
\dontshow{
summary(cvCheck(m, dd, ncpus = 1, n.init = 10))
summary(cvCheck(m, dd, ncpus = 4, n.init = 10))
}

}
