% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_df.R
\name{dfNuclear}
\alias{dfNuclear}
\title{Compute the degree of freedom for nuclear regularized regression}
\usage{
dfNuclear(B.LS, B.lambda, lambda, lambda2 = 0, tol = 1e-12)
}
\references{
Zhou et al. Regularized Matrix Regression
}

