% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lava_modelsearch.R
\name{extendModel}
\alias{extendModel}
\alias{extendModel.lvm}
\title{Automatic extension of the lvm}
\usage{
extendModel(x, ...)

\method{extendModel}{lvm}(x, type, covariance = TRUE, data, alpha = 0.05,
  method = "holm", display.warnings = TRUE, trace = TRUE, ...)
}
\arguments{
\item{x}{a lvm model}

\item{...}{additional arguments to be passed to the \code{estimate} method.}

\item{type}{should all links be added to the latent variable model ("all") or only thoses relevant according to a score ("modelsearch") or a LR ("modelsearchLR") test.}

\item{covariance}{should covariance links be considered?}

\item{data}{the dataset used to perform the model search}

\item{alpha}{the significance threshold for retaining a new link}

\item{method}{the method used to adjust the p.values for multiple comparisons in the modelsearch}

\item{display.warnings}{should warnings be display? May occur when dealing with categorical variables or when fitting an extended model.}

\item{trace}{should the execution of the modelsearch be traced?}
}
\value{
a latent variable model
}
\examples{
mSim <- lvm()
regression(mSim) <- c(y1,y2,y3)~u
regression(mSim) <- u~x1+x2
categorical(mSim,labels=c("A","B","C")) <- "x2"
latent(mSim) <- ~u
covariance(mSim) <- y1~y2
df <- sim(mSim, 1e2)

m <- lvm(c(y1,y2,y3)~u)
latent(m) <- ~u
addvar(m) <- ~x1+x2 

coef(extendModel(m, type = "all"))
coef(extendModel(m, type = "all", covariance = FALSE, display.warnings = FALSE))

coef(extendModel(m, type = "modelsearch", data = df))
coef(extendModel(m, type = "modelsearchLR", data = df))
}

