% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lava_extract.R
\name{extractCoef}
\alias{coefCov}
\alias{coefCov.lvm}
\alias{coefCov.lvmfit}
\alias{coefExtra}
\alias{coefExtra.lvm}
\alias{coefExtra.lvmfit}
\alias{coefIntercept}
\alias{coefIntercept.lvm}
\alias{coefIntercept.lvmfit}
\alias{coefRef}
\alias{coefRef.lvmfit}
\alias{coefReg}
\alias{coefReg.lvm}
\alias{coefReg.lvmfit}
\alias{coefType}
\alias{coefType.lvm}
\alias{coefType.lvmfit}
\alias{coefVar}
\alias{coefVar.lvm}
\alias{coefVar.lvmfit}
\alias{extractCoef}
\title{Extract the specific coefficient names or positions in a LVM}
\usage{
coefType(x, ...)

\method{coefType}{lvm}(x, ...)

\method{coefType}{lvmfit}(x, level = 9, ...)

coefCov(x, ...)

\method{coefCov}{lvm}(x, value = FALSE, keep.var = FALSE, ...)

\method{coefCov}{lvmfit}(x, value = FALSE, keep.var = FALSE, ...)

coefExtra(x, ...)

\method{coefExtra}{lvm}(x, value = FALSE, ...)

\method{coefExtra}{lvmfit}(x, value = FALSE, ...)

coefIntercept(x, ...)

\method{coefIntercept}{lvm}(x, value = FALSE, ...)

\method{coefIntercept}{lvmfit}(x, value = FALSE, ...)

coefRef(x, ...)

\method{coefRef}{lvmfit}(x, value = FALSE, ...)

coefReg(x, ...)

\method{coefReg}{lvm}(x, value = FALSE, ...)

\method{coefReg}{lvmfit}(x, value = FALSE, ...)

coefVar(x, ...)

\method{coefVar}{lvm}(x, value = FALSE, ...)

\method{coefVar}{lvmfit}(x, value = FALSE, ...)
}
\arguments{
\item{x}{a lvm model or a fitted lvm model}

\item{value}{should the name of the coefficient be returned? Else return the coefficients}

\item{keep.var}{should the variance parameters be output?}
}
\examples{
## regression
m <- lvm(Y~X1+X2)
e <- estimate(m, sim(m, 1e2))

coefType(m)
coefType(e)
coefType(e, level = -1)

coefCov(m)
coefCov(m, value = TRUE)

coefCov(m, keep.var = TRUE)
coefCov(m, value = TRUE, keep.var = TRUE)

coefIntercept(m)
coefIntercept(m, value = TRUE)

coefReg(m)
coefReg(m, value = TRUE)

## LVM
m <- lvm()
regression(m) <- c(y1,y2,y3)~u
regression(m) <- u~x
latent(m) <- ~u
covariance(m) <- y1~y2

e <- estimate(m, sim(m, 1e2))

coefType(m)
coefType(e)
coefType(e, level = -1)

coefCov(m)
coefCov(m, value = TRUE)#' 

coefCov(m, keep.var = TRUE)
coefCov(m, value = TRUE, keep.var = TRUE)

coefExtra(m)

categorical(m, K = 3) <- "X1"
coefExtra(m)
coefExtra(m, value = TRUE)

coefIntercept(m)
coefIntercept(m, value = TRUE)
coefIntercept(e)

coefReg(e, value = TRUE)
coefReg(e, level = -1, value = TRUE)

}

