% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lava_reducedMoment.R
\name{gaussianLP_score.lvm}
\alias{gaussianLP_score.lvm}
\title{Compute the score for a reduced lvm model}
\usage{
gaussianLP_score.lvm(x, p, data, indiv = FALSE, ...)
}
\details{
this function assumes that the external parameters in p are at the end of the vector
}

