% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_extractPath.R
\name{getPath}
\alias{getPath}
\alias{getPath.plvmfit}
\title{Get the regularization path}
\usage{
getPath(x, ...)

\method{getPath}{plvmfit}(x, names = NULL, getCoef, getLambda,
  only.breakpoints = FALSE, increasing = TRUE, order = "lambda1",
  row = NULL)
}
\arguments{
\item{x}{a penalized lvm model}

\item{names}{names of the columns to extract}

\item{getCoef}{names of the coefficients to extract. Can also be \code{"coef0"} or \code{"coefn0"} to extract the number of 0 or non-0 coefficients}

\item{getLambda}{names of the penalization parameter to extract.}

\item{only.breakpoints}{should the path be extracted only at the breakpoints}

\item{increasing}{should the path be extracted by increasing value of regularization parameter}

\item{order}{the regularization parameter to consider when ordering the points along the path}

\item{row}{the index at which the path should be extracted}
}
\description{
Extract the regularization path from a lvm object
}

