% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_glmPath.R
\name{glmPath}
\alias{glmPath}
\title{Compute the regularization path for a penalized regression}
\usage{
glmPath(beta0, objective, hessian, gradient, indexPenalty, indexNuisance, sd.X,
  base.lambda1, lambda2, test.penalty1, control, iter.max = length(beta0) * 3)
}

