% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_proxOperator.R
\name{init.proxOperator}
\alias{init.proxOperator}
\title{Initialise a proximal operator}
\usage{
init.proxOperator(lambda1, lambda2, group.coef, lambdaN, nrow, ncol,
  regularizationPath)
}
\arguments{
\item{lambda1}{penalization parameter for the (group) lasso penalty}

\item{lambda2}{penalization parameter for the ridge penalty}

\item{group.coef}{group of coefficients where a common lasso penalty should be applied}

\item{lambdaN}{penalization parameter for the nuclear norm penalty}

\item{nrow}{the number of rows of the matrix of coefficients}

\item{ncol}{the number of columns of the matrix of coefficients}

\item{regularizationPath}{Will the regularization path be computed? In such a case always include the lasso penalty.}
}
\value{
a list containing the proximal operator and the penalty funtion
}
\description{
Generate the proximal operator for the requested penalty
}
\examples{
lava.penalty:::init.proxOperator(lambda1 = 1, lambda2 = NULL, group.coef = 1:5, lambdaN = NULL, nrow = NA, ncol = NA, regularizationPath = FALSE)

}

