% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_penalize.R
\name{initGroup.lvm}
\alias{initGroup.lvm}
\title{Reshape information for building the V matrices}
\usage{
initGroup.lvm(x, links, group)
}
\arguments{
\item{x}{a \code{lvm}-object}

\item{links}{the name of the penalized links}

\item{group}{define the groups of the penalty}
}
\description{
Return matrices containg the endogenous, exogenous and group index corresponding to each penlized link. Take care of categorical variables.
}
\examples{
## no category
m <- lvm(Y~X1+X2+X3)
dt <- initGroup.lvm(m, links = c("Y~X1","Y~X2"))
dt

## categories
categorical(m, labels = c("A","B","C")) <- "X1"
dt <- initGroup.lvm(m, links = c("Y~X1","Y~X2"))
dt
categorical(m, K=2, labels = 1:2) <- "X2"
dt <- initGroup.lvm(m, links = c("Y~X1","Y~X2"))
dt

regression(m) <- Z~X1+X3+X5+eta
latent(m) <- ~eta
dt <- initGroup.lvm(m, links = c("Y~X1","Y~X2","Z~X1"))
dt
}
