% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_initializer.R
\name{initializeData}
\alias{initializeData}
\alias{initializeData.plvm}
\alias{initializeData.pmultigroup}
\title{Scale the dataset and convert categorical variables to indicators}
\usage{
initializeData(x, ...)

\method{initializeData}{plvm}(x, data, method.center = "mean",
  method.scale = "sd")

\method{initializeData}{pmultigroup}(x, method.center = "mean",
  method.scale = "sd")
}
\arguments{
\item{x}{a penalized lvm model}

\item{data}{a data.frame containing the data}

\item{method.center}{function used to center the data}

\item{method.scale}{function used to scale the data}
}
\description{
Scale the dataset and convert categorical variables to indicators
}
