% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_initializer.R
\name{initializeOperator}
\alias{initializeOperator}
\title{Initialise a proximal operator}
\usage{
initializeOperator(lambda1, index.penalty1, index.model1, lambda2,
  index.penalty2, index.model2, lambdaG, index.penaltyG, index.modelG, lambdaN,
  index.penaltyN, nrow, ncol, equivariance, constrain.variance, index.variance)
}
\arguments{
\item{lambda1}{lasso penalization parameter.}

\item{index.penalty1}{a list containing the position of lasso penalized parameters in the vector of parameters.}

\item{index.model1}{the index of the model relative to each lasso penalized parameter}

\item{lambda2}{ridge penalization parameter.}

\item{index.penalty2}{a list containing the position of ridge penalized parameters in the vector of parameters.}

\item{index.model2}{the index of the model relative to each ridge penalized parameter}

\item{lambdaG}{group lasso penalization parameter.}

\item{index.penaltyG}{a list containing the position of group lasso parameters in the vector of parameters.}

\item{index.modelG}{the index of the model relative to each group lasso penalized parameter}

\item{lambdaN}{nuclear norm penalization parameter.}

\item{index.penaltyN}{a list containing the position of the parameters corresponding to each image in the vector of parameters.}

\item{nrow}{a vector containing the number of rows of each image.}

\item{ncol}{a vector containing the number of columnss of each image.}

\item{equivariance}{should the lambda parameter be multiplied with the first variance parameter?}

\item{constrain.variance}{should the variance parameters be exponential trans}

\item{index.variance}{the position of the variance parameters in coef}
}
\value{
a list containing the proximal operator and the penalty funtion
}
\description{
Generate the proximal operator for the requested penalty
}
\examples{
lavaPenalty:::initializeOperator(lasso = TRUE, ridge = FALSE, groupLasso = FALSE, nuclearNorm = FALSE)

}
