% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_estimate.R
\name{initializer.lvm}
\alias{initializer.lvm}
\title{Find an initial solution for the model}
\usage{
initializer.lvm(x, data, names.coef, n.coef, penalty, regPath, ...)
}
\arguments{
\item{x}{a penalized lvm model}

\item{data}{a data.frame containing the data}

\item{n.coef}{number of coefficients}

\item{penalty}{parameter of the penalty}

\item{regPath}{parameter of the penalisation path}
}

