% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_hook.R
\name{initializer.plvm}
\alias{initializer.plvm}
\title{Intialization of the parameter of the plvm}
\usage{
\method{initializer}{plvm}(x, data, ...)
}
\arguments{
\item{x}{a penalized lvm model}

\item{data}{a data.frame containing the data}
}
\description{
Compute the coefficients of the non and the completely regularized latent variable model
}
\examples{
m <- lvm()
regression(m) <- c(y1,y2,y3)~u+x1+x2+x3
regression(m) <- u~x1+x2
latent(m) <- ~u
covariance(m) <- y1 ~ y2

pm <- penalize(m)
initializer(pm, data = sim(pm, 1e2))

}

