% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_penalize.R
\name{initPenalty}
\alias{initPenalty}
\alias{initPenaltyL12}
\alias{initPenaltNuclear}
\title{Initialize penalty terms}
\usage{
initPenaltyL12()

initPenaltNuclear()
}
\description{
Initialise lasso, ridge, group lasso, and nuclear norm penalties
}
\details{
initPenaltyL12 contains the specification of the lasso, ridge, group lasso penalties
initPenaltNuclear contains the specification of the nuclear norm penalty
}
