% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_kill.R
\name{kill.penalty}
\alias{kill.penalty}
\alias{lavaPenalty.remove.hook}
\title{Remove penalty after removing a variable}
\usage{
lavaPenalty.remove.hook(x, var, ...)
}
\arguments{
\item{x}{\code{plvm}-object}

\item{var}{the names of the variables that should be removed}

\item{...}{not used}

\item{value}{the names of the link that should be removed}
}
\description{
Remove penalty after removing a variable
}
\examples{
m <- lvm()
m <- regression(m, x=paste0("x",1:10),y="y")
pm <- penalize(m)
kill(pm) <- "x7"

res <- cancel(pm, y~x1+x2)
res

res <- cancel(pm, y~x1+x2+x3)
res
}
