% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_options.R
\name{lava.penalty.options}
\alias{lava.penalty.options}
\title{Set global options for lava.penalty}
\usage{
lava.penalty.options(..., type = "lava")
}
\arguments{
\item{...}{Arguments}

\item{type}{which set of parameters to update: \code{"lava"}, \code{"proxGrad"}, or \code{"EPSODE"}.}
}
\examples{

## lava
lava.penalty.options()
lava.penalty.options(trace = TRUE)

## lava.penalty
lava.penalty.options(type = "proxGrad")
lava.penalty.options(trace = 2, type = "proxGrad")
}

