% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lava_penalty-package.R
\docType{package}
\name{lava.penalty}
\alias{lava.penalty}
\alias{lava.penalty-package}
\title{lava.penalty: Regularization tools for latent variables models}
\description{
add-on to the lava package allowing regularized linear models and latent variable models.
}

