% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_hook.R
\name{lavaPenalty.multigroup.hook}
\alias{lavaPenalty.multigroup.hook}
\title{Hook to estimate a multigroup penalized lvm model}
\usage{
lavaPenalty.multigroup.hook(mg, x, ...)
}
\description{
Merge the penalty of the different models into one
}
