% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_options.R
\name{lavaPenalty.options}
\alias{lavaPenalty.options}
\title{Set global options for lavaPenalty}
\usage{
lavaPenalty.options(..., type = "lava")
}
\arguments{
\item{...}{Arguments}

\item{type}{which set of parameters to update: \code{"lava"}, \code{"proxGrad"}, or \code{"EPSODE"}.}
}
\description{
Extract and set global parameters of lavaPenalty.
}
\examples{

## lava
lavaPenalty.options()
lavaPenalty.options(trace = TRUE)

## lavaPenalty
lavaPenalty.options(type = "proxGrad")
lavaPenalty.options(trace = 2, type = "proxGrad")
}
