% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LavaPenalty-package.R
\docType{package}
\name{lavaPenalty}
\alias{lavaPenalty}
\alias{lavaPenalty-package}
\title{lavaPenalty: Regularization tools for latent variables models}
\description{
add-on to the lava package allowing regularized linear models and latent variable models.
}
