% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{LCSseq}
\alias{LCSseq}
\title{Common substring sequence}
\usage{
LCSseq(x)
}
\arguments{
\item{x}{a list of character}
}
\description{
get the common substring sequence among a vector of strings
}
\examples{
\dontrun{
LCSseq <- lavaPenalty:::LCSseq
LCSseq(list("ad","a","ad"))
}
}
