% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lava_extract.R
\name{loadings}
\alias{loadings}
\alias{loadings.lvm.missing}
\alias{loadings.lvmfit}
\title{Extract the summary table for the loadings}
\usage{
\method{loadings}{lvmfit}(x, col = NULL)

\method{loadings}{lvm.missing}(x, col = NULL)
}
\arguments{
\item{x}{a lvm model}

\item{value}{should the name of the coefficient be returned? Else return the coefficients}
}
\examples{
m <- lvm()
regression(m) <- c(y1,y2,y3)~u
regression(m) <- u~x
latent(m) <- ~u
df <- sim(m,1e2)

fit <- estimate(m, df)

loadings(fit) 
loadings(fit, col = "P-value") 
}

