% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lava_reducedObject.R
\name{lp<-}
\alias{lp<-}
\title{Update of the linear predictor}
\usage{
lp(x, ...) <- value
}
\arguments{
\item{x}{\code{lvm}-object}

\item{lp}{the name of the linear predictors to be updated}

\item{value}{the value that will be allocated}
}
\description{
Update one or several linear predictors
}
\examples{

## regresssion
m <- lvm()
m <- regression(m, x=paste0("x",1:10),y="y", reduce = TRUE)
vars(m)

newLP <- lp(m, type = NULL)[[1]]
newLP$link <- newLP$link[1:3]
newLP$con <- newLP$con[1:3]
newLP$x <- newLP$x[1:3]

lp(m, lp = 1) <- newLP
lp(m, type = NULL)

}

