% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_penalize.R
\name{lvm2plvm}
\alias{lvm2plvm}
\title{Conversion to a penalized latent variable model}
\usage{
lvm2plvm(x)
}
\arguments{
\item{x}{\code{lvm}-object}
}
\description{
Conversion to a penalized latent variable model
}

