% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lava_modelsearchLR.R, R/Lava_modelsearchMax.R
\name{modelsearchLR}
\alias{modelsearchLR}
\alias{modelsearchMax}
\title{Model searching using a likelihood ratio test}
\usage{
modelsearchLR(x, data = NULL, restricted, link, directive,
  ls.LVMargs = NULL, method.p.adjust = "bonferroni",
  display.warnings = TRUE, trace = 1, ...)

modelsearchMax(x, data = NULL, restricted, link, directive,
  method.max = "boot-wild", alpha = 0.05, ls.LVMargs = NULL,
  method.p.adjust = "bonferroni", n.sim = 10000, conditional = NULL,
  mu.conditional = NULL, iid.conditional = NULL, export.iid = 1,
  trace = 1, ncpus = 1, ...)
}
\arguments{
\item{x}{a lvm model}

\item{display.warnings}{should the warnings encountered after the fit of an extended model be displayed?}

\item{...}{additional arguments to be passed to the \code{estimate} method.}

\item{na.omit}{do not export the results for links where the extended model has not converged}

\item{silent}{should the execution of the modelsearch be traced?}

\item{x}{a lvm model}

\item{na.omit}{do not export the results for links where the extended model has not converged}

\item{display.warnings}{should the warnings encountered after the fit of an extended model be displayed?}

\item{silent}{should the execution of the modelsearch be traced?}

\item{...}{additional arguments to be passed to the \code{estimate} method.}
}
\value{
an list

an object of class lvmfit
}
\description{
Model searching using a likelihood ratio test

Model searching using the max statistic to retain or not a link
}
\seealso{
\code{link{modelsearch2}}

\code{link{modelsearch2}}
}
