% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_glmPath.R
\name{nextNode_lvm}
\alias{nextNode_lvm}
\title{Find the next value of the regularization parameter}
\usage{
nextNode_lvm(hessian, gradient, gradientPen, beta, lambda1, lambda2,
  indexPenalty, indexNuisance)
}

