% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_glmPath.R
\name{orthoData_glmPath}
\alias{orthoData_glmPath}
\title{Orthogonalize the non-penalized variables relatively to the penalized variables for a regression model}
\usage{
orthoData_glmPath(model, name.Y, allCoef, penaltyCoef, data)
}

