% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{p2j}
\alias{p2j}
\alias{p2j.dgCMatrix}
\title{Extract the column index from a sparse matrix}
\usage{
p2j(x, ...)

\method{p2j}{dgCMatrix}(x)
}
\arguments{
\item{x}{sparse matrix}
}
\description{
Extract the column index from a sparse matrix
}
\examples{
p2j <- lavaSearch2:::p2j
initVcoef.lvm <- lavaSearch2:::initVcoef.lvm

m <- lvm(Y~X1+X2+X3)
V <- initVcoef.lvm(m, link = coefReg(m), group = 1:length(coefReg(m)))
p2j(V)
p2j(cbind(V,0,V))
p2j(cbind(0,V,0,V))

V2 <- rbind(1,V)
p2j(V2)
p2j(cbind(0,V2,0,V2))

}
