% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_penalty.R
\name{penaltyExtract}
\alias{penalty}
\alias{penalty.lvm}
\alias{penalty.lvmfit}
\alias{penalty.penaltyL12}
\alias{penalty.penaltyNuclear}
\alias{penaltyExtract}
\title{Extract the penalty from a lvm object}
\usage{
penalty(x, ...)

\method{penalty}{lvm}(x, type = NULL, nuclear = FALSE,
  lambdaPerCoef = FALSE, add.names = TRUE, ...)

\method{penalty}{lvmfit}(x, type, ...)

\method{penalty}{penaltyL12}(x, type = NULL, no.lasso = FALSE,
  no.ridge = FALSE, no.group = FALSE, index.group = NULL, ...)

\method{penalty}{penaltyNuclear}(x, type, group = NULL, keep.list = FALSE,
  ...)
}
\arguments{
\item{x}{a lvm model}

\item{type}{the information about the penalty to be returned.}

\item{nuclear}{should informations be extracted for the nuclear penalty}

\item{lambdaPerCoef}{when a regularization parameter is extracted should it expanded to match the parameters of the LVM}

\item{add.names}{when a regularization parameter is extracted should the name of the parameters be added in the output. Only active if lambdaPerCoef is TRUE.}

\item{group}{the group of penalty parameter to be returned}

\item{keep.list}{should the format of the results be always a list}
}
\description{
Extract the penalty from a lvm object
}
\examples{
set.seed(10)
n <- 300
formula.lvm <- as.formula(paste0("Y~",paste(paste0("X",1:12), collapse = "+")))
mSim <- lvm(formula.lvm)
df.data <- sim(mSim,n)

pm <- penalize(mSim, c("Y~X1","Y~X4", "Y~X10"))
penalty(pm, type = "link")
penalty(pm, type = c("link","group"))

pm <- penalize(pm, value = paste0("Y~X",5:6), group = 1, add = TRUE)
penalty(pm, type = "link") # all parameters
penalty(pm, type = "link", group = 0) # individually penalized parameters
penalty(pm, type = "link", group = 1) # group 1 penalization
penalty(pm, type = "V")
penalty(pm, type = "V", group = 0)

penalty(pm, type = "lambda1")
penalty(pm, type = "lambda1", lambdaPerCoef = TRUE)
penalty(pm, type = "lambda1", lambdaPerCoef = TRUE, add.names = FALSE)
penalty(pm, type = c("lambda1","lambda2"), lambdaPerCoef = TRUE)

penalty(pm, type = "link") <- c("Y~X1" , "Y~X4")

}

