% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_penalty.R
\name{penaltyUpdate}
\alias{penaltyUpdate}
\alias{penalty<-}
\alias{penalty<-.plvm}
\alias{penalty<-.penaltyL12}
\alias{penalty<-.penaltyNuclear}
\title{Update the penalty term}
\usage{
penalty(x, ...) <- value

\method{penalty}{plvm}(x, nuclear = FALSE, ...) <- value

\method{penalty}{penaltyL12}(x, type, lambda, add, ...) <- value

\method{penalty}{penaltyNuclear}(x, type = "link", ...) <- value
}
\arguments{
\item{x}{a lvm.penalty object}

\item{nuclear}{should informations be extracted for the nuclear penalty}

\item{value}{the value to be attributed}

\item{type}{the type of penalty information to be updated}
}
\description{
Update the penalty term of a lvm.penalty object
}
\examples{
set.seed(10)
n <- 300
formula.lvm <- as.formula(paste0("Y~",paste(paste0("X",1:12), collapse = "+")))
mSim <- lvm(formula.lvm)
df.data <- sim(mSim,n)

pm <- penalize(mSim, c("Y~X1","Y~X4", "Y~X10"))
pm

}
