% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_estimate.R
\name{prepareData.lvm}
\alias{prepareData.lvm}
\title{Prepare the data for the estimate function
 
Scale the data, update the penalty term according to the presence of factors. If regularizationPath=1 (i.e. glmPath) orthogonolize the data}
\usage{
prepareData.lvm(x, data, method.center = "mean", method.scale = "sd",
  penalty)
}
\arguments{
\item{x}{a penalized lvm model}

\item{data}{a data.frame containing the data}

\item{penalty}{parameter of the penalty}
}

