% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_estimate.R
\name{prepareData.plvm}
\alias{prepareData.plvm}
\title{Prepare the data for the estimate function}
\usage{
prepareData.plvm(x, data, method.center = "mean", method.scale = "sd")
}
\arguments{
\item{x}{a penalized lvm model}

\item{data}{a data.frame containing the data}

\item{method.center}{function used to center the data}

\item{method.scale}{function used to scale the data}
}
\description{
Scale the data, update the penalty term according to the presence of factors.
}

