% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_print.R
\name{print.penalize}
\alias{print.penalize}
\alias{print.penaltyL12}
\alias{print.penaltyNuclear}
\alias{print.plvm}
\alias{print.plvmfit}
\title{Display the content of a penalized object}
\usage{
\method{print}{penaltyL12}(x, ...)

\method{print}{penaltyNuclear}(x, ...)

\method{print}{plvm}(x, ...)

\method{print}{plvmfit}(x, level = 2, labels = FALSE, coef, lambda = NULL,
  only.breakpoints = NULL, ...)
}
\arguments{
\item{x}{a plvm object}

\item{...}{additional arguments to be passed to lower level functions.}
}
\description{
Display the content of a penalized object
}

