% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_print.R
\name{print.plvmfit}
\alias{print.plvmfit}
\title{Display the content of a plvmfit object}
\usage{
\method{print}{plvmfit}(x, level = 2, labels = FALSE, coef = "penalized",
  lambda = "abs", only.breakpoints = TRUE, ...)
}
\arguments{
\item{x}{a plvmfit object}
}

