% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lava_reducedHook.R
\name{procdata.lvm.reduced}
\alias{procdata.lvm.reduced}
\title{Preparation of data}
\usage{
\method{procdata}{lvm.reduced}(x, data, ...)
}
\description{
add the LPs in the first and second empirical moment so that LP is not converted to a latent variable
}

