% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_proxOperator.R
\name{proximalOperators}
\alias{proximalOperators}
\alias{proxL1}
\alias{prox}
\alias{proxL2}
\alias{proxE2}
\alias{proxNuclear}
\alias{proxL2}
\alias{proxE2}
\alias{proxNuclear}
\title{Proximal operators}
\usage{
proxL1(x, step, lambda)

proxL2(x, step, lambda)

proxE2(x, step, lambda)

proxNuclear(x, step, lambda, nrow, ncol)
}
\arguments{
\item{x}{value of the coefficients}

\item{step}{value of the step size. Should be between 0 and the inverse of the Lipschitw constant of grad f}

\item{lambda}{value of the penalization parameter}

\item{nrow}{the number of rows of the matrix of coefficients}

\item{ncol}{the number of columns of the matrix of coefficients}
}
\value{
the updated vector of coefficients
}
\description{
Proximal operators for the lasso, ridge, group lasso and nuclear norm penalties.
}
