% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lava_reducedObject.R
\name{reduce}
\alias{reduce}
\title{Reduce latent variable model}
\usage{
reduce(x, ...)
}
\arguments{
\item{x}{\code{lvm}-object}

\item{endo}{the endogeneous variables for which the related exogeneous variables should be aggregated}

\item{rm.exo}{should the exogeneous variables be remove from the object}
}
\description{
Aggregate exogeneous variables into a linear predictor
}
\examples{
## regression
m <- lvm()
m <- regression(m, x=paste0("x",1:10),y="y")

m
reduce(m)
coef(reduce(m))
reduce(m, rm.exo = FALSE)

## lvm
m <- lvm()
m <- regression(m, x=paste0("x",1:10),y="y1")
m <- regression(m, x=paste0("x",51:150),y="y2")
covariance(m) <- y1~y2

m
reduce(m)
coef(reduce(m))
reduce(m, rm.exo = FALSE)

## with penalty
m <- lvm()
m <- regression(m, x=paste0("x",1:10),y="y")
pm <- penalize(m)
reduce(pm)
reduce(pm, link = y~x1+x2+x3)

}

