% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{select.response}
\alias{select.response}
\title{Response variable of a formula}
\usage{
select.response(formula)
}
\arguments{
\item{formula}{a formula}
}
\description{
Return the reponse variable contained in the formula
}
\examples{
select.response(Y1~X1+X2)
}

