% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lava_cv.R
\name{summary.cvlvm}
\alias{summary.cvlvm}
\title{Summary function associated with cvCheck}
\usage{
summary.cvlvm(object, threshold = NULL, ...)
}
\arguments{
\item{object}{the output of cvCheck}

\item{threshold}{threshold used to cluster the convergence points (height in hclust)}
}
\description{
Summary function associated with cvCheck
}
