% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduced_object.R
\name{cancelLP}
\alias{cancelLP}
\alias{cancelLP.lvm.reduced}
\title{Remove variable from the linear predictor}
\usage{
\method{cancelLP}{lvm.reduced}(x, link = NULL, lp = NULL, expar = TRUE,
  simplify = TRUE)
}
\arguments{
\item{x}{\code{lvm}-object}

\item{link, value}{the name of the links that should be removed}

\item{lp}{the name of the linear predictors that should be removed}

\item{expar}{should the external parameters be also removed}

\item{simplify}{should the class \code{lvm.reduced} be removed is from the \code{lvm}-object if it contains no LP.}

\item{restaure}{should the link be kept while removing the linear predictor}

\item{...}{argument passed to \code{cancelLP.lvm.reduced}}
}
\description{
Remove one or several variable from the linear predictor
}
\examples{

}

