% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_cancelPenalty.R
\name{cancelPenalty}
\alias{cancelPenalty}
\alias{cancelPenalty.plvm}
\alias{cancelPenalty<-}
\alias{cancelPenalty<-.penaltyL12}
\alias{cancelPenalty<-.plvm}
\title{Remove penalty from a penalized latent variable model}
\usage{
cancelPenalty(x, ...)

cancelPenalty(x, ...) <- value

\method{cancelPenalty}{plvm}(x, simplify = TRUE, link)

\method{cancelPenalty}{plvm}(x, clean = TRUE, ...) <- value

\method{cancelPenalty}{penaltyL12}(x, extraParameter) <- value
}
\arguments{
\item{x}{\code{plvm}-object}

\item{value}{the penalty that should be removed}

\item{simplify}{if the object contain no more penalty should it be converted to a non penalized lvm}

\item{link}{the penalty that should be removed}
}
\description{
Remove one or several penalties from a penalized latent variable model
}
\examples{

## lasso
m <- lvm()
m <- regression(m, x=paste0("x",1:10),y="y")
pm <- penalize(m)

cancelPenalty(pm, link = "y~x5")
cancelPenalty(pm) <- y~x1
cancelPenalty(pm) <- c("y~x2","y~x3")
cancelPenalty(pm) <- penalty(pm)
pm

## group lasso
m <- regression(m, x=paste0("x",1:10),y="y")
categorical(m, K = 3) <- ~x1
pm <- penalize(m)
cancelPenalty(pm) <- "x1:0|1"
pm
cancelPenalty(pm) <- "x1:1|2"
pm
cancelPenalty(pm) <- penalty(pm, type = "link")
pm


}

