% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_EPSODE.R
\name{EPSODE}
\alias{EPSODE}
\title{Perform the generic Path Algorithm for a LVM}
\usage{
EPSODE(start, objective, gradient, hessian, V, lambda2, index.penalty2,
  equivariance, constrain.variance, index.variance, control)
}
\arguments{
\item{start}{the starting value}

\item{objective}{likelihood given by lava. Used to adjust the step parameter when using backtracking}

\item{gradient}{first derivative of the likelihood given by lava.}

\item{hessian}{second derivative of the likelihood given by lava. Only used to estimate the step parameter of the algorithm when step = NULL}

\item{V}{matrix that left multiply beta to define the penalization (identity corresponds to a standard lasso penalty)}

\item{lambda2}{ridge penalization parameter}

\item{index.penalty2}{parameters to which ridge penalization is applied}

\item{equivariance}{should the lambda parameter be multiplied with the first variance parameter?}

\item{constrain.variance}{should the variance parameters be log transformed?}

\item{index.variance}{the position of the variance parameters in start}

\item{control}{settings for the EPSODE algorithm. See lava.options.}
}
\description{
Perform the generic Path Algorithm for a LVM
}
\details{
Does not work for an unknown variance matrix since the log-likelihood of the regression and variance parameter is not jointly convex
(according to ??)
}
\references{
Zhou 2014 - A generic Path Algorithm for Regularized Statistical Estimation
}
