% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lava_link.R
\name{findNewLink}
\alias{findNewLink}
\alias{findNewLink.lvm}
\title{Find the new possible links between variables (copied from lava::modelsearch)}
\usage{
findNewLink(x, ...)

\method{findNewLink}{lvm}(x, rm.latent = FALSE, rm.endo = FALSE,
  output = "names")
}
\arguments{
\item{x}{a lvm model}

\item{...}{not used}

\item{rm.latent}{ignore links between latent variables}

\item{rm.endo}{ignore links between endogenous variables}

\item{output}{return the names of the variables to link ("names") or their position ("index")}
}
\description{
Find the new possible links between variables (copied from lava::modelsearch)
}
\examples{
\dontrun{
findNewLink <- lava.penalty:::findNewLink

m <- lvm()
regression(m) <- c(y1,y2,y3)~u
findNewLink(m, rm.endo = FALSE)
findNewLink(m, rm.endo = TRUE)

regression(m) <- u~x1+x2
latent(m) <- ~u

findNewLink(m, rm.endo = FALSE)
findNewLink(m, rm.endo = TRUE)
findNewLink(m, rm.endo = TRUE, output = "index")
}
}

