% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduced_object.R
\name{reduce}
\alias{reduce}
\alias{reduce.plvm}
\title{Reduce latent variable model}
\usage{
\method{reduce}{plvm}(object, link = NULL, ...)
}
\arguments{
\item{...}{other arguments to be passed to \code{lava.reduce::reduce.lvm}}

\item{x}{\code{lvm}-object}

\item{endo}{the endogeneous variables for which the related exogeneous variables should be aggregated}

\item{rm.exo}{should the exogeneous variables be remove from the object}
}
\description{
Aggregate exogeneous variables into a linear predictor
}
\examples{
 
m <- lvm()
m <- regression(m, x=paste0("x",1:10),y="y")
pm <- penalize(m)
reduce(pm)
reduce(pm, link = y~x1+x2+x3)

}

