% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_initializer.R
\name{initializeOperator}
\alias{initializeOperator}
\title{Initialise a proximal operator}
\usage{
initializeOperator(lambda1, lambda2, lambdaG, index.penaltyG, lambdaN,
  index.penaltyN, nrow, ncol, constrain.lambda, constrain.variance,
  index.variance)
}
\arguments{
\item{lambda1}{lasso penalization parameter.}

\item{lambda2}{ridge penalization parameter.}

\item{lambdaG}{group lasso penalization parameter.}

\item{index.penaltyG}{a list containing the position of each group of parameters in the vector of parameters.}

\item{lambdaN}{nuclear norm penalization parameter.}

\item{index.penaltyN}{a list containing the position of the parameters corresponding to each image in the vector of parameters.}

\item{nrow}{a vector containing the number of rows of each image.}

\item{ncol}{a vector containing the number of columnss of each image.}

\item{constrain.lambda}{should the lambda parameter be normalized by the total variance?}

\item{constrain.variance}{should the variance parameters be exponential trans}
}
\value{
a list containing the proximal operator and the penalty funtion
}
\description{
Generate the proximal operator for the requested penalty
}
\examples{
lava.penalty:::initializeOperator(lasso = TRUE, ridge = FALSE, groupLasso = FALSE, nuclearNorm = FALSE)

}

