% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_optims.R
\name{initPenalty}
\alias{initPenalty}
\title{Initialise the penalty}
\usage{
initPenalty(pen, penNuclear, start = NULL, name.coef = NULL,
  regPath = NULL)
}
\arguments{
\item{pen}{definition of the penalty}

\item{penNuclear}{definition of the nuclear norm penalty}

\item{start}{initialization value of the parameters}

\item{name.coef}{names of the parameters}

\item{regPath}{control parameters for the regularization path}
}
\description{
Update the penalty according to the initialization of estimate
}

