% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_optims.R, R/Penalty_penalize.R
\name{initPenalty}
\alias{initPenaltNuclear}
\alias{initPenalty}
\alias{initPenaltyL12}
\title{Initialise the penalty}
\usage{
initPenalty(pen, penNuclear, start = NULL, name.coef = NULL,
  regPath = NULL)

initPenaltyL12()

initPenaltNuclear()
}
\arguments{
\item{pen}{definition of the penalty}

\item{penNuclear}{definition of the nuclear norm penalty}

\item{start}{initialization value of the parameters}

\item{name.coef}{names of the parameters}

\item{regPath}{control parameters for the regularization path}
}
\description{
Update the penalty according to the initialization of estimate

Initialise lasso, ridge, group lasso, and nuclear norm penalties
}
\details{
initPenaltyL12 contains the specification of the lasso, ridge, group lasso penalties
initPenaltNuclear contains the specification of the nuclear norm penalty
}

