% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{initVar_link}
\alias{initVar_link}
\title{Normalize var1 and var2}
\usage{
initVar_link(var1, var2, repVar1 = FALSE, format = "list")
}
\arguments{
\item{var1}{a character indicating the endogeneous variable or a formula}

\item{var2}{an optional character indicating the exogeneous variable}

\item{repVar1}{should var1 be duplicated to match var2 length. Only active if format = "list".}

\item{format}{should the name of the variable be return (format = "list"), a vector of character formula ("txt.formula") or a list of formula ("formula")}
}
\description{
Convert var1 and var2 from formula or covariance to character
}
\examples{
lava.penalty:::initVar_link(var1 = a~b)
lava.penalty:::initVar_link(var1 = a ~ b)
lava.penalty:::initVar_link(var1 = a ~ b+c+d*e, format = "list")
lava.penalty:::initVar_link(var1 = a ~ b+c+d*e, format = "txt.formula")
lava.penalty:::initVar_link(var1 = a ~ b+c+d*e, format = "formula")

lava.penalty:::initVar_link(var1 = "a,b")
lava.penalty:::initVar_link(var1 = "a", var2 = "b")

lava.penalty:::initVar_link(var1 = Y~X1+X2)
lava.penalty:::initVar_link(var1 = Y~X1+X2, repVar1 = TRUE)
lava.penalty:::initVar_link(var1 = Y~X1+X2, format = "formula")
lava.penalty:::initVar_link(var1 = Y~X1+X2, format = "txt.formula")
}

