% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lava_modelsearch.R
\name{modelsearchLR}
\alias{modelsearchLR}
\alias{modelsearchLR.lvmfit}
\title{Model searching using a likelihood ratio test}
\usage{
modelsearchLR(object, ...)

\method{modelsearchLR}{lvmfit}(object, na.omit = FALSE,
  display.warnings = FALSE, silent = FALSE, ...)
}
\arguments{
\item{...}{additional arguments to be passed to the \code{estimate} method.}

\item{na.omit}{do not export the results for links where the extended model has not converged}

\item{display.warnings}{should the warnings encountered after the fit of an extended model be displayed?}

\item{silent}{should the execution of the modelsearch be traced?}

\item{x}{a lvm model}
}
\value{
an object of class lvmfit
}
\description{
Model searching using a likelihood ratio test
}
\examples{
mSim <- lvm()
regression(mSim) <- c(y1,y2,y3)~u
regression(mSim) <- u~x1+x2
categorical(mSim,labels=c("A","B","C")) <- "x2"
latent(mSim) <- ~u
covariance(mSim) <- y1~y2
df <- sim(mSim, 1e2)

m <- lvm(c(y1,y2,y3)~u)
latent(m) <- ~u
addvar(m) <- ~x1+x2 

fit <- estimate(m, data = df)
modelsearchLR(fit)
modelsearchLR(fit, na.omit = TRUE)

}

