% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_optims.R
\name{optim.Nuisance.plvm}
\alias{optim.Nuisance.plvm}
\title{Estimate nuisance parameter}
\usage{
optim.Nuisance.plvm(x, data, coefEstimated, coefNuisance, coefPenalty, control,
  ...)
}
\arguments{
\item{x}{a penalized lvm model}

\item{data}{a data.frame containing the data}

\item{coefEstimated}{the estimated mean coefficients}

\item{coefNuisance}{the name of the nuisance parameters to be estimated}

\item{coefPenalty}{the name of the penalized parameters}

\item{control}{control arguments to be passed to estimate.lvm}
}
\description{
Use a constrained LVM to estimate nuisance parameters
}

