% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_ISTA.R
\name{proxGrad}
\alias{proxGrad}
\title{step of a proximal gradient algorithm}
\usage{
proxGrad(start, proxOperator, hessian, gradient, objective, iter.max, abs.tol,
  rel.tol, trace, step = lava.options()$proxGrad$step,
  BT.n = lava.options()$proxGrad$BT.n,
  BT.eta = lava.options()$proxGrad$BT.eta,
  force.descent = lava.options()$proxGrad$force.descent,
  method = lava.options()$proxGrad$method)
}
\arguments{
\item{start}{initial values for the parameters}

\item{proxOperator}{proximal operator corresponding to the penalization applied to the log likelihood}

\item{hessian}{second derivative of the likelihood given by lava. Only used to estimate the step parameter of the algorithm when step = NULL}

\item{gradient}{first derivative of the likelihood given by lava.}

\item{objective}{likelihood given by lava. Used to adjust the step parameter when using backtracking}

\item{iter.max}{maximum number of iterations}

\item{abs.tol}{convergence is the difference in likelihood between two consecutive steps is below this threshold}

\item{rel.tol}{convergence is the relative difference  in likelihood between two consecutive steps is below this threshold}

\item{trace}{should the convergence diagnostics be displayed at each step}

\item{step}{maximun step for the proximal gradient algorithm. 
If NULL the step is estimated using the inverse of the maximal eigenvalue of the hessian (in absolute value) and re-estimated at each step
Otherwise backtracking is used.}

\item{BT.n}{number of backtracking steps}

\item{BT.eta}{multiplicative factor for the step}

\item{lambda1}{L1 penalization parameter}

\item{lambda2}{L2 penalization parameter}

\item{group.lambda1}{group to which each parameter belongs. 0 mean individual lasso otherwise parameters are groupes according to their group.lambda1 value}

\item{fast}{type of iteration
0 correspond to the ISTA step as described in Bech 2009
1 correspond to the FISTA step as described in Bech 2009
2 correspond to the Monotone APG as described in Li 2015
3 correspond to the Nesterov step as described in Simon 2013}
}
\references{
Bech and Teboulle - 2009 A Fast Iterative Shrinkage-Thresholding Algorithm
Li 2015 - Accelerated Proximal Gradient Methods for Nonconvex Programming
Simon 2013 - A sparse group Lasso
}

