% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalty_ISTA.R
\name{proxGrad}
\alias{proxGrad}
\title{Proximal gradient algorithm}
\usage{
proxGrad(start, proxOperator, hessian, gradient, objective,
  control = lava.options()$proxGrad)
}
\arguments{
\item{start}{initial values for the parameters}

\item{proxOperator}{proximal operator corresponding to the penalization applied to the log likelihood}

\item{hessian}{second derivative of the likelihood given by lava. Only used to estimate the step parameter of the algorithm when step = NULL}

\item{gradient}{first derivative of the likelihood given by lava.}

\item{objective}{likelihood given by lava. Used to adjust the step parameter when using backtracking}

\item{control}{settings for the proximal gradient algorithm. See lava.options.}
}
\description{
Estimate parameters using a proximal gradient algorithm
}
\references{
Bech and Teboulle - 2009 A Fast Iterative Shrinkage-Thresholding Algorithm
Li 2015 - Accelerated Proximal Gradient Methods for Nonconvex Programming
Simon 2013 - A sparse group Lasso
}
