% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduced_object.R
\name{reduce.penalty}
\alias{lava.penalty.reduce.hook}
\alias{reduce.penalty}
\title{Aggregate exogeneous variables associated with penalties}
\usage{
lava.penalty.reduce.hook(x, link = NULL, ...)
}
\arguments{
\item{x}{\code{lvm}-object}

\item{link}{the links that should be included in the linear predictor}

\item{...}{other arguments to be passed to \code{lava.reduce::reduce.lvm}}
}
\description{
Aggregate exogeneous variables associated with penalties into a linear predictor
}
\examples{
 
m <- lvm()
m <- regression(m, x=paste0("x",1:10),y="y")
pm <- penalize(m)
res <- reduce(pm)
reduce(pm, link = y~x1+x2+x3)

}

