% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{renameFactor}
\alias{renameFactor}
\title{Find the categorical variable matching the name of the coefficient in lava}
\usage{
renameFactor(var, ls.levels, data, sep = "")
}
\examples{
renameFactor("x2C", ls.levels = list(x1 = 1:5, x2 = c("A","B","C")))

}

